+++ src/commands/mod.rs
@@ -1,3 +1,5 @@
 pub mod read;
 pub mod write;
 pub mod label;
+pub mod tree;
+pub mod trueid;